+++ src/Scriptbook/Private/Get-AnsiColoredString.ps1
@@ -1,5 +1,8 @@
-function Get-AnsiColoredString([string]$String, [ValidateNotNull()][int]$Color, [switch]$NotSupported)
+function Get-AnsiColoredString
 {
+    [Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSAvoidGlobalVars", "")]
+    param([string]$String, [ValidateNotNull()][int]$Color, [switch]$NotSupported)
+
     # ref: https://en.wikipedia.org/wiki/ANSI_escape_code for color codes 
     # ref: https://duffney.io/usingansiescapesequencespowershell/
     if ($Global:ScriptbookSimpleHost -or $NotSupported.IsPresent)