+++ src/Scriptbook/Private/Invoke-AlwaysActions.ps1
@@ -1,6 +1,7 @@
 function Invoke-AlwaysActions
 {
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSShouldProcess", "")]
+    [Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSUseSingularNouns", "")]
     [CmdletBinding(SupportsShouldProcess = $true)]
     param (
         $Actions,