+++ src/Scriptbook/Public/Utilities/Test-IsSecureStringStorageObject.ps1
@@ -1,0 +1,4 @@
+function Test-IsSecureStringStorageObject([ValidateNotNull()]$Object)
+{
+    return ($Object -is [SecureStringStorage])
+}